# backend/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vacancy API",
        "description": "REST API для управления вакансиями",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local Docker server"
        }
    ],
    "paths": {
        "/api/vacancies": {
            "get": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Список вакансий",
                "operationId": "9fd1df96854f4d5e7b57844adda9c475",
                "responses": {
                    "200": {
                        "description": "Список вакансий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vacancy"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Создать вакансию",
                "operationId": "e50935cda3d16b2e61c82b398fe6e286",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VacancyCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vacancy"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/vacancies/{id}": {
            "get": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Получить вакансию",
                "operationId": "c1af0904f1ed226fc3eadc124742eb76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вакансия",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vacancy"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Vacancy": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "VacancyCreate": {
                "required": [
                    "title",
                    "description",
                    "salary"
                ],
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Vacancies",
            "description": "Vacancies"
        }
    ]
}